Moved compile config to new file compile-config.ts
