+++ kipper/core/src/compiler/index.ts
@@ -8,6 +8,7 @@ export * from "./ast/";
 export * from "./analysis/";
 export * from "./target-presets/";
 export * from "./optimiser/";
+export * from "./compile-config";
 export * from "./compiler";
 export * from "./program-ctx";
 export * from "./runtime-built-ins";
